Atualizado Validacao dos dados

+++ src/controllers/BandasController.js
@@ -1,24 +1,34 @@
+const { validationResult } = require('express-validator');
 const Banda = require('../models/Bandas');
 
 class BandasController {
   static adicionar = async (req, res) => {
     try {
-      const novaBanda = {
-        nome: req.body.nome,
-        ano_formacao: req.body.ano_formacao,
-        ativa: req.body.ativa,
-        genero: req.body.genero,
-        regiao: req.body.regiao,
-        descricao: req.body.descricao,
-      };
+      const validacao = validationResult(req);
 
-      await Banda.create(novaBanda);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const novaBanda = {
+          nome: req.body.nome,
+          ano_formacao: req.body.ano_formacao,
+          ativa: req.body.ativa,
+          genero: req.body.genero,
+          regiao: req.body.regiao,
+          descricao: req.body.descricao,
+        };
 
-      res.status(201).json({
-        mensagem: `Banda ${novaBanda.nome} adicionado com sucesso!`,
-        dados: novaBanda,
-        status: 201,
-      });
+        await Banda.create(novaBanda);
+
+        res.status(201).json({
+          mensagem: `Banda ${novaBanda.nome} adicionado com sucesso!`,
+          dados: novaBanda,
+          status: 201,
+        });
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -30,9 +40,18 @@ class BandasController {
 
   static exibirTodos = async (req, res) => {
     try {
-      const bandas = await Banda.find();
+      const validacao = validationResult(req);
 
-      res.status(200).json(bandas);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const bandas = await Banda.find();
+
+        res.status(200).json(bandas);
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -44,11 +63,27 @@ class BandasController {
 
   static exibirUm = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      const banda = await Banda.findById(id);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      res.status(200).json(banda);
+        const banda = await Banda.findById(id);
+
+        if (banda) {
+          res.status(200).json(banda);
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe banda com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -60,24 +95,42 @@ class BandasController {
 
   static atualizar = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      const atualizacaoBanda = {
-        nome: req.body.nome,
-        ano_formacao: req.body.ano_formacao,
-        ativa: req.body.ativa,
-        genero: req.body.genero,
-        regiao: req.body.regiao,
-        descricao: req.body.descricao,
-      };
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      await Banda.findByIdAndUpdate(id, atualizacaoBanda);
+        const banda = await Banda.findById(id);
 
-      res.status(200).json({
-        mensagem: `Banda de id ${id} foi atualiado com sucesso!`,
-        dados: atualizacaoBanda,
-        status: 200,
-      });
+        if (banda) {
+          const atualizacaoBanda = {
+            nome: req.body.nome,
+            ano_formacao: req.body.ano_formacao,
+            ativa: req.body.ativa,
+            genero: req.body.genero,
+            regiao: req.body.regiao,
+            descricao: req.body.descricao,
+          };
+
+          await Banda.findByIdAndUpdate(id, atualizacaoBanda);
+
+          res.status(200).json({
+            mensagem: `Banda de id ${id} foi atualiado com sucesso!`,
+            dados: atualizacaoBanda,
+            status: 200,
+          });
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe banda com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -89,14 +142,32 @@ class BandasController {
 
   static deletar = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      await Banda.findByIdAndDelete(id);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      res.status(200).json({
-        mensagem: `Banda de id ${id} foi deletado com sucesso!`,
-        status: 200,
-      });
+        const banda = await Banda.findById(id);
+
+        if (banda) {
+          await Banda.findByIdAndDelete(id);
+
+          res.status(200).json({
+            mensagem: `Banda de id ${id} foi deletado com sucesso!`,
+            status: 200,
+          });
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe banda com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
